requirements/common.txt
Django>=1.8,<1.9
psycopg2>=2.6,<2.7
django-bootstrap3>=4.8,<4.9
django-grappelli>=2.7.2,<2.8
PyYAML>=3.11,<4
django-registration-redux>=1.2,<1.3
django-crispy-forms>=1.5.2,<1.6
django-guardian>=1.3.2,<1.4
requests>=2.9.1,<3.0.0 # Hera
pypandoc>=1.1.2,<1.2.0
bleach>=1.4.2,<1.5.0
bleach-whitelist
django-floppyforms